intro-to-vector-db/pyproject.toml
[project]
name = "intro-to-vector-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.13",
    "langchain-pinecone>=0.2.5",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.0",
]
